+++ src/point.rs
@@ -313,6 +313,10 @@ impl<T: Unit> Point2<T> {
         glam::Vec2 =>
         #[doc = "Extend with z-component to [`Point3`]."]
         pub fn extend(self, z: T::Scalar) -> Point3<T>;
+        #[doc = "Replace the x-component with a new value."]
+        pub fn with_x(self, x: T::Scalar) -> Self;
+        #[doc = "Replace the y-component with a new value."]
+        pub fn with_y(self, y: T::Scalar) -> Self;
     );
 
     point_interface!(Point2, Vector2);
@@ -383,6 +387,12 @@ impl<T: Unit> Point3<T> {
         pub fn extend(self, w: T::Scalar) -> Point4<T>;
         #[doc = "Truncate to [`Point2`]."]
         pub fn truncate(self) -> Point2<T>;
+        #[doc = "Replace the x-component with a new value."]
+        pub fn with_x(self, x: T::Scalar) -> Self;
+        #[doc = "Replace the y-component with a new value."]
+        pub fn with_y(self, y: T::Scalar) -> Self;
+        #[doc = "Replace the z-component with a new value."]
+        pub fn with_z(self, z: T::Scalar) -> Self;
     );
 
     point_interface!(Point3, Vector3);
@@ -458,6 +468,14 @@ impl<T: Unit> Point4<T> {
         glam::Vec4 =>
         #[doc = "Truncate to [`Point3`]."]
         pub fn truncate(self) -> Point3<T>;
+        #[doc = "Replace the x-component with a new value."]
+        pub fn with_x(self, x: T::Scalar) -> Self;
+        #[doc = "Replace the y-component with a new value."]
+        pub fn with_y(self, y: T::Scalar) -> Self;
+        #[doc = "Replace the z-component with a new value."]
+        pub fn with_z(self, z: T::Scalar) -> Self;
+        #[doc = "Replace the w-component with a new value."]
+        pub fn with_w(self, w: T::Scalar) -> Self;
     );
 
     point_interface!(Point4, Vector4);