+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `fract_gl()`
+- `with_x()`, `with_y()`, etc., as well as `with_width()` and `with_height()`
+  for size types.
 ### Breaking changes
 - Bumped dependency on `glam` to 0.27.0.
 - Minimum Supported Rust Version bumped to 1.68.2 for `impl From<bool> for